+++ bigquery/src/lib.rs
@@ -1,6 +1,5 @@
 extern crate core;
 
-pub mod arrow;
 pub mod client;
 pub mod grpc;
 pub mod http;
